+++ Epoch.net/Constants.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Epoch.net
+{
+    public static class Constants
+    {
+        public static DateTime UnixEpoch = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
+    }
+}